Code Generation!
Macros weren't quite cutting it and I was having to copy-paste a lot of
almost-identical code all over the place anyway, so I decided to bite
the bullet and do some code generation instead.

It's fairly simplistic right now, since I wanted something that I could
get up and running fairly easily. I do, however, now have a
machine-readable description of all 290 instructions in core SPIR-V to
build on.

This generates the Instruction definition, some code for extracting some
common data from an instruction and the parser.

It also removes the module stuff, as I wasn't happy with the design
still, and adds a license pre-amble to all the files.