Add a RawModule interface for working with SPIR-V modules
Implement a simple, relatively low-level interface for dealing SPIR-V
modules. RawModule will just parse in the entire module and build some
maps to make finding defintions and uses of ids a bit faster.

Adds an example that dumps the variables in an SPIR-V module.

Fixes a bug in the instruction description, Source was missing a
parameter.